Add reopen tests for FileZipInput
