+++ subprojects/files/src/test/groovy/org/gradle/api/internal/file/archive/impl/StreamZipInputTest.groovy
@@ -20,14 +20,9 @@ import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.junit.Rule
 import spock.lang.Specification
 
-import java.util.zip.ZipEntry
-import java.util.zip.ZipOutputStream
-
-
-class StreamZipInputTest extends Specification {
+class StreamZipInputTest extends Specification implements ZipFileFixture {
     @Rule
     TestNameTestDirectoryProvider temporaryFolder = new TestNameTestDirectoryProvider(getClass())
-    private static final ZIP_ENTRY_CONTENT = "foo"
 
     def "cannot read from a zip entry stream a second time"() {
         def zipInput = new StreamZipInput(makeZip("foo.zip").newInputStream())
@@ -69,28 +64,5 @@ class StreamZipInputTest extends Specification {
 
         cleanup:
         zipInput?.close()
-    }
-
-    private static byte[] readAllBytes(InputStream is) {
-        ByteArrayOutputStream baos = new ByteArrayOutputStream()
-        int b = is.read()
-        while(b != -1) {
-            baos.write(b)
-            b = is.read()
-        }
-        return baos.toByteArray()
-    }
-
-    private File makeZip(String zipFileName) {
-        def zipFile = temporaryFolder.file(zipFileName)
-        def file = temporaryFolder.file("foo.txt")
-        file.text = ZIP_ENTRY_CONTENT
-        def zipStream = new ZipOutputStream(zipFile.newOutputStream())
-        def zipEntry = new ZipEntry(file.name)
-        zipEntry.setSize(file.size())
-        zipStream.putNextEntry(zipEntry)
-        zipStream.write(file.bytes)
-        zipStream.close()
-        return zipFile
     }
 }
